Add PlaceType component and update usage in PlaceTypes and PlaceModal
